Move caffe_function from functions to links
