+++ chainer/links/caffe/__init__.py
@@ -1,5 +1,4 @@
 from chainer.links.caffe import caffe_function
 
 
-# for backward compatibility
 CaffeFunction = caffe_function.CaffeFunction
